YLW-3314 chore: renamed update files
