+++ MusicSchool/Requests/Category/UpdateCategory.cs
@@ -1,0 +1,6 @@
+﻿namespace MusicSchool.Requests.Category;
+
+public class UpdateCategory
+{
+    public required string NewCategoryName { get; set; }
+}
