--- MusicSchool/Requests/CategoryPut.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicSchool.Requests;
-
-public class CategoryPut
-{
-    public required string NewCategoryName { get; set; }
-}
