+++ MusicSchool/Requests/Instrument/UpdateInstrumentPut.cs
@@ -1,6 +1,6 @@
-namespace MusicSchool.Requests;
+namespace MusicSchool.Requests.Instrument;
 
-public class InstrumentPut
+public class UpdateInstrumentPut
 {
     public required string NewInstrumentName { get; set; }
     public required int NewCategoryId { get; set; }
