+++ MusicSchool/Requests/Student/UpdateStudentInstrumentsPatch.cs
@@ -1,0 +1,6 @@
+﻿namespace MusicSchool.Requests.Student;
+
+public class UpdateStudentInstrumentsPatch
+{
+    public IEnumerable<int> NewInstrumentIds { get; set; } = default!;
+}
