--- MusicSchool/Requests/StudentInstrumentPatch.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicSchool.Requests;
-
-public class StudentInstrumentPatch
-{
-    public IEnumerable<int> NewInstrumentIds { get; set; } = default!;
-}
